REVIT-184051 Help with open source cleanup.
