+++ Data converting/ButtonAvailability.cs
@@ -1,0 +1,15 @@
+﻿using Autodesk.Revit.Attributes;
+using Autodesk.Revit.DB;
+using Autodesk.Revit.UI;
+
+namespace RevitToDatabase
+{
+    [Transaction(TransactionMode.Manual)]
+    public class ButtonAvailability : IExternalCommandAvailability
+    {
+        public bool IsCommandAvailable(UIApplication applicationData, CategorySet selectedCategories)
+        {
+            return true;
+        }
+    }
+}
