+++ Data converting/packages.config
@@ -1,15 +1,16 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="BouncyCastle" version="1.8.3.1" targetFramework="net472" />
-  <package id="DynamicLanguageRuntime" version="1.2.3" targetFramework="net472" />
+  <package id="DynamicLanguageRuntime" version="1.3.0" targetFramework="net472" />
   <package id="Google.Protobuf" version="3.6.1" targetFramework="net472" />
-  <package id="IronPython" version="2.7.10" targetFramework="net472" />
+  <package id="IronPython" version="2.7.11" targetFramework="net472" />
   <package id="K4os.Compression.LZ4" version="1.1.11" targetFramework="net472" />
   <package id="K4os.Compression.LZ4.Streams" version="1.1.11" targetFramework="net472" />
   <package id="K4os.Hash.xxHash" version="1.0.6" targetFramework="net472" />
   <package id="MySql.Data" version="8.0.21" targetFramework="net472" />
   <package id="SSH.NET" version="2016.1.0" targetFramework="net472" />
   <package id="System.Buffers" version="4.5.0" targetFramework="net472" />
+  <package id="System.Collections" version="4.3.0" targetFramework="net472" />
   <package id="System.Memory" version="4.5.3" targetFramework="net472" />
   <package id="System.Numerics.Vectors" version="4.4.0" targetFramework="net472" />
   <package id="System.Runtime.CompilerServices.Unsafe" version="4.6.0" targetFramework="net472" />
